more work on saving recipes to users
added logic to save recipes to users schema so we can later access them and show their saved recipes.
My current plan is to get the document ID's, save em to a list in the user schema, when we want to see our saved recipes we'll just find all recipes by id that are saved to the user's schema.
- Typescript is throwing a FIT! really in line 34: if I set line20 to userdeets:String; then I get an error `Type 'String' is missing the following properties from type 'StringConstructor'` can't set it to `any` type either, similar error. so I have 'bypassed' (put off till later) the error by hardcoding the userdeets value. 'magically' the error goes away. Seems to be a bit too strongly typed to me, why can't I initalize a string without properties, I mean what other language do you have to specify PROPERTIES for a STRING!...oh well we'll get it figured out, I hope.