reworked backend recipe
reworked code to where it actually works now ;)
ref lines 8-26, found working solution in reddit of all places, crawled stackoverflow and google for a good while, visited many blogs and finally wound up here: https://www.reddit.com/r/learnjavascript/comments/55mjr7/mongoose_and_inserting_subdocuments_to_a_user/
   Specifically top comment: "First you need to User.findOne({_id:req.body.id}, function(err,doc){ If(err)res.sendStatus(500).send('database error' ).end(); If(!doc){res.sendStatus(404).send('user was not found').end() }else{ doc.logsHeld.push({logName: req.body.logNameInput}); doc.markModified('logsHeld') doc.save(); res.sendStatus(200).send('log saved').end(); } })

First you have to pull a user from the database. Once Mongoose returns the document from the "findOne" method you can modify the document. You need to tell the database hey I modified something make sure you save it. Which is why mongoose has the portion to 'markModified' then you have to send it to the database to be saved. So you end with doc.save().

But you also need to respond to the client or user or whomever by sending a A Okay and ideally some message so that a notification system on the front end can signal the user that it happened. Or you can return other stuff. Or you can go to the next middleware. Or you can or you can or you can............ Do anything your heart desires wooo programming.

Also sorry for formatting on my phone and it can never get the code right."

updated getSavedRecipes...guess technically I am just reading the user here (so this is duplicate code), especially since I switched up the recipe sub schema to save the whole recipe obj to the user's doc.