it's ugly don't look
I didn't want to do this...but I had to due to time constraints,
this way has a funnier? output than my og method for ingredients, now foreach item searched EVERY card has EVERY ingredient AND measurement on it :) makes for one REALLY long list! oh well, this is a bug for a later date